implement nav-bar with main routing
- implement navbar components and main routing components
- change some css styles
- make sperate page for listing products and add route for it
