--- host/src/Components/NavBar.jsx
@@ -1,10 +1,0 @@
-import React from 'react'
-
-
-export default function NavBar() {
-    return (
-        <div className="navbar bg-base-100 shadow">
-            <a className="btn btn-ghost text-xl">daisyUI</a>
-        </div>
-    )
-}
