+++ host/src/Components/NavBar/SearchBar.jsx
@@ -1,0 +1,9 @@
+import React from 'react'
+
+export default function SearchBar() {
+    return (
+        <div className="form-control">
+            <input type="text" placeholder="Search" className="input input-bordered w-24 md:w-auto" />
+        </div>
+    )
+}
