+++ host/src/Components/NavBar/TabList.jsx
@@ -1,0 +1,14 @@
+import React from 'react'
+import { Link } from 'react-router-dom'
+
+export default function TabList({ routes }) {
+    return (
+        <ul className='tabs' role="tablist">
+            {routes.map((route, index) => (
+                <li key={index} className='tab text-lg'>
+                    <Link to={route.path}>{route.label}</Link>
+                </li>
+            ))}
+        </ul>
+    );
+}
